Themes: Refactor Checkbox, Switch, and Radio (#1204)
