+++ src/components/checkbox/checkbox-base.ts
@@ -118,6 +118,7 @@ export class IgcCheckboxBaseComponent extends FormAssociatedRequiredMixin(
 
   protected handleClick() {
     this.checked = !this.checked;
+    this.focused = false;
     this.emitEvent('igcChange', { detail: this.checked });
   }
