+++ src/components/checkbox/themes/dark/checkbox/checkbox.fluent.scss
@@ -7,3 +7,9 @@ $theme: $fluent;
 :host {
     @include css-vars-from-theme(diff(light.$base, $theme), 'ig-checkbox');
 }
+
+:host([invalid]:hover) {
+    [part~='indicator'] {
+        stroke: color('error', 500);
+    }
+}
