+++ src/components/checkbox/themes/dark/switch/_themes.scss
@@ -1,7 +1,6 @@
 @use 'styles/utilities' as *;
 @use 'igniteui-theming/sass/themes/schemas/components/dark/switch' as *;
 
-$base: digest-schema($base-dark-switch);
 $material: digest-schema($dark-material-switch);
 $bootstrap: digest-schema($dark-bootstrap-switch);
 $fluent: digest-schema($dark-fluent-switch);
