+++ src/components/checkbox/themes/dark/switch/switch.indigo.scss
@@ -1,8 +1,9 @@
 @use 'styles/utilities' as *;
 @use 'themes' as *;
+@use '../../light/switch/themes' as light;
 
 $theme: $indigo;
 
 :host {
-    @include css-vars-from-theme(diff($base, $theme), 'ig-switch');
+    @include css-vars-from-theme(diff(light.$base, $theme), 'ig-switch');
 }
