--- src/components/checkbox/themes/dark/switch/switch.shared.scss
@@ -1,6 +1,0 @@
-@use 'styles/utilities' as *;
-@use 'themes' as *;
-
-:host {
-    @include css-vars-from-theme($base, 'ig-switch');
-}
