+++ src/components/checkbox/themes/shared/checkbox/checkbox.fluent.scss
@@ -33,7 +33,6 @@ $disabled-color: var-get($theme, 'disabled-color');
 }
 
 [part~='indicator'] {
-    stroke: var-get($theme, 'tick-color');
     stroke-width: 1.5;
     transform: scale(.8);
 }
@@ -51,8 +50,7 @@ $disabled-color: var-get($theme, 'disabled-color');
     }
 }
 
-:host(:hover) [part='control checked'],
-:host(:focus-within) [part='control checked'] {
+:host(:hover) [part='control checked'] {
     &::after {
         background: $checked-hover-color;
         box-shadow: inset 0 0 0 rem(1px) $checked-hover-color;
@@ -88,7 +86,7 @@ $disabled-color: var-get($theme, 'disabled-color');
 }
 
 :host([indeterminate]:hover) [part~='control'],
-:host([indeterminate]:focus-within) [part~='control'] {
+:host([indeterminate]:focus-visible) [part~='control'] {
     &::before {
         background: $checked-hover-color;
     }
@@ -118,7 +116,7 @@ $disabled-color: var-get($theme, 'disabled-color');
     }
 }
 
-:host([invalid]:focus-within) {
+:host([invalid]:focus-visible) {
     [part~='control'] {
         &::after {
             box-shadow: inset 0 0 0 rem(1px) $invalid-hover-color;
@@ -167,7 +165,7 @@ $disabled-color: var-get($theme, 'disabled-color');
         }
     }
 
-    &:hover {
+    &:focus-within {
         [part~='control'] {
             &::before {
                 background: $invalid-hover-color;
@@ -179,17 +177,18 @@ $disabled-color: var-get($theme, 'disabled-color');
             }
         }
     }
+}
 
-    &:focus-within {
-        [part~='control'] {
-            &::before {
-                background: $invalid-hover-color;
-            }
-
-            &::after {
-                box-shadow: inset 0 0 0 rem(1px) $invalid-hover-color;
-                background: transparent;
-            }
+:host(:hover:not([disabled])[invalid][indeterminate]),
+:host(:hover:not(:disabled)[invalid][indeterminate]) {
+    [part~='control'] {
+        &::before {
+            background: $invalid-hover-color;
+        }
+ 
+        &::after {
+            box-shadow: inset 0 0 0 rem(1px) $invalid-hover-color;
+            background: transparent;
         }
     }
 }
@@ -212,7 +211,7 @@ $disabled-color: var-get($theme, 'disabled-color');
         position: absolute;
         top: -$focus-outline-offset-top;
         left: -$focus-outline-offset-left;
-        box-shadow: 0 0 0 rem(1px) color(gray, 700);
+        box-shadow: 0 0 0 rem(1px) var-get($theme, 'focus-outline-color');
         width: calc(100% + (#{$focus-outline-offset-left} * 2));
         height: calc(100% + (#{$focus-outline-offset-top} * 2));
     }
