+++ src/components/checkbox/themes/shared/checkbox/checkbox.indigo.scss
@@ -21,12 +21,10 @@ $error-color-hover: var-get($theme, 'error-color-hover');
     width: var(--size, rem(16px));
     height: var(--size, rem(16px));
     min-width: var(--size, rem(16px));
-    transition: background .15s ease-in;
 
     &::after {
         border-radius: var-get($theme, 'border-radius');
         border: rem(2px) solid $empty-color;
-        transition: border-color .15s ease-in;
     }
 }
 
@@ -36,7 +34,6 @@ $error-color-hover: var-get($theme, 'error-color-hover');
 }
 
 [part~='indicator'] {
-    stroke: var-get($theme, 'tick-color');
     transform: scale(.8);
 }
 
@@ -63,11 +60,15 @@ $error-color-hover: var-get($theme, 'error-color-hover');
     }
 }
 
-:host(:focus-within) {
+:host(:focus-visible),
+[part='base focused'] {
     [part~='control'] {
         box-shadow: 0 0 0 rem(3px) var-get($theme, 'focus-outline-color');
     }
+}
 
+:host(:focus-visible),
+[part='base checked focused'] {
     [part='control checked'] {
         box-shadow: 0 0 0 rem(3px) var-get($theme, 'focus-outline-color-focused');
     }
@@ -75,7 +76,7 @@ $error-color-hover: var-get($theme, 'error-color-hover');
 
 :host([indeterminate]) {
     [part~='indicator'] {
-        transform: rotate(45deg) translateX(-.125rem) scale(.8);
+        transform: rotate(45deg) translateX(-.1rem) scale(.8);
     }
 
     [part~='control'] {
@@ -95,7 +96,7 @@ $error-color-hover: var-get($theme, 'error-color-hover');
     }
 }
 
-:host([indeterminate]:focus-within) {
+:host([indeterminate]:focus-visible) {
     [part~='control'] {
         box-shadow: 0 0 0 rem(3px) var-get($theme, 'focus-outline-color-focused');
     }
@@ -165,16 +166,16 @@ $error-color-hover: var-get($theme, 'error-color-hover');
     }
 }
 
-:host(:focus-within:not([disabled])[invalid]),
-:host(:focus-within:not(:disabled)[invalid]) {
+:host(:focus-visible:not([disabled])[invalid]),
+:host(:focus-visible:not(:disabled)[invalid]) {
     [part~='control'],
     [part='control checked'] {
         box-shadow: 0 0 0 rem(3px) var-get($theme, 'focus-outline-color-error');
     }
 }
 
-:host(:focus-within:not([disabled])[invalid][indeterminate]),
-:host(:focus-within:not(:disabled)[invalid][indeterminate]) {
+:host(:focus-visible:not([disabled])[invalid][indeterminate]),
+:host(:focus-visible:not(:disabled)[invalid][indeterminate]) {
     [part~='control'] {
         box-shadow: 0 0 0 rem(3px) var-get($theme, 'focus-outline-color-error');
     }
