+++ src/components/checkbox/themes/shared/checkbox/checkbox.material.scss
@@ -7,8 +7,8 @@ $theme: $material;
     @include type-style('subtitle-1');
 }
 
-[part~='indicator'] {
-    stroke: var-get($theme, 'tick-color');
+label {
+    gap: rem(2px);
 }
 
 [part~='control'] {
@@ -64,7 +64,7 @@ $theme: $material;
     }
 }
 
-[part='base']:focus-within,
+[part='base']:focus-visible,
 [part='base focused'] {
     &:hover [part~='control'],
     [part~='control'] {
@@ -75,7 +75,7 @@ $theme: $material;
     }
 }
 
-[part='base checked']:focus-within,
+[part='base checked']:focus-visible,
 [part='base checked focused'] {
     &:hover [part~='control'],
     [part~='control'] {
@@ -94,18 +94,11 @@ $theme: $material;
         }
     }
 
-    [part='base']:hover {
-        [part~='control']::before {
-            background: var-get($theme, 'fill-color');
-            opacity: .12;
-        }
-    }
-
-    [part='base']:focus-within,
+    [part='base']:hover,
+    [part='base']:focus-visible,
     [part='base focused'] {
         [part~='control']::before {
             background: var-get($theme, 'fill-color');
-            opacity: .12;
         }
     }
 }
@@ -126,7 +119,7 @@ $theme: $material;
 
     [part~='base']:hover {
         [part~='control']::before {
-            background: var-get($theme, 'error-color-hover');
+            background: var-get($theme, 'error-color');
         }
     }
