+++ src/components/checkbox/themes/switch-themes.ts
@@ -6,13 +6,12 @@ import { styles as bootstrapDark } from './dark/switch/switch.bootstrap.css.js';
 import { styles as fluentDark } from './dark/switch/switch.fluent.css.js';
 import { styles as indigoDark } from './dark/switch/switch.indigo.css.js';
 import { styles as materialDark } from './dark/switch/switch.material.css.js';
-import { styles as sharedDark } from './dark/switch/switch.shared.css.js';
 // Light Overrides
 import { styles as bootstrapLight } from './light/switch/switch.bootstrap.css.js';
 import { styles as fluentLight } from './light/switch/switch.fluent.css.js';
 import { styles as indigoLight } from './light/switch/switch.indigo.css.js';
 import { styles as materialLight } from './light/switch/switch.material.css.js';
-import { styles as sharedLight } from './light/switch/switch.shared.css.js';
+import { styles as shared } from './light/switch/switch.shared.css.js';
 // Shared Styles
 import { styles as bootstrap } from './shared/switch/switch.bootstrap.css.js';
 import { styles as fluent } from './shared/switch/switch.fluent.css.js';
@@ -21,7 +20,7 @@ import { styles as material } from './shared/switch/switch.material.css.js';
 
 const light = {
   shared: css`
-    ${sharedLight}
+    ${shared}
   `,
   bootstrap: css`
     ${bootstrap} ${bootstrapLight}
@@ -39,7 +38,7 @@ const light = {
 
 const dark = {
   shared: css`
-    ${sharedDark}
+    ${shared}
   `,
   bootstrap: css`
     ${bootstrap} ${bootstrapDark}
