+++ src/components/radio/themes/dark/_themes.scss
@@ -1,7 +1,6 @@
 @use 'styles/utilities' as *;
 @use 'igniteui-theming/sass/themes/schemas/components/dark/radio' as *;
 
-$base: digest-schema($base-dark-radio);
 $material: digest-schema($dark-material-radio);
 $bootstrap: digest-schema($dark-bootstrap-radio);
 $fluent: digest-schema($dark-fluent-radio);
