+++ src/components/radio/themes/dark/radio.bootstrap.scss
@@ -1,8 +1,16 @@
 @use 'styles/utilities' as *;
 @use 'themes' as *;
+@use '../light/themes' as light;
 
 $theme: $bootstrap;
 
 :host {
-    @include css-vars-from-theme(diff($base, $theme), 'ig-radio');
+    @include css-vars-from-theme(diff(light.$base, $theme), 'ig-radio');
+}
+
+:host([disabled]),
+:host(:disabled) {
+    [part='control'] {
+        background: color('surface')
+    }
 }
