--- src/components/radio/themes/dark/radio.shared.scss
@@ -1,6 +1,0 @@
-@use 'styles/utilities' as *;
-@use 'themes' as *;
-
-:host {
-    @include css-vars-from-theme($base, 'ig-radio');
-}
