+++ src/components/radio/themes/themes.ts
@@ -6,13 +6,12 @@ import { styles as bootstrapDark } from './dark/radio.bootstrap.css.js';
 import { styles as fluentDark } from './dark/radio.fluent.css.js';
 import { styles as indigoDark } from './dark/radio.indigo.css.js';
 import { styles as materialDark } from './dark/radio.material.css.js';
-import { styles as sharedDark } from './dark/radio.shared.css.js';
 // Light Overrides
 import { styles as bootstrapLight } from './light/radio.bootstrap.css.js';
 import { styles as fluentLight } from './light/radio.fluent.css.js';
 import { styles as indigoLight } from './light/radio.indigo.css.js';
 import { styles as materialLight } from './light/radio.material.css.js';
-import { styles as sharedLight } from './light/radio.shared.css.js';
+import { styles as shared } from './light/radio.shared.css.js';
 // Shared Styles
 import { styles as bootstrap } from './shared/radio.bootstrap.css.js';
 import { styles as fluent } from './shared/radio.fluent.css.js';
@@ -21,7 +20,7 @@ import { styles as material } from './shared/radio.material.css.js';
 
 const light = {
   shared: css`
-    ${sharedLight}
+    ${shared}
   `,
   bootstrap: css`
     ${bootstrap} ${bootstrapLight}
@@ -39,7 +38,7 @@ const light = {
 
 const dark = {
   shared: css`
-    ${sharedDark}
+    ${shared}
   `,
   bootstrap: css`
     ${bootstrap} ${bootstrapDark}
